Fix deleteAndBan action dryRun and small refactoring
 * Fix the deleteAndBan action to actually work in dry run
 * Handle EditMessage properly
 * Rename state type package
