+++ filters/types/checkNeventsState/gen.go
@@ -1,3 +1,3 @@
-package state
+package checkNeventsState
 
 //go:generate protoc --proto_path=. --go_out=. --go_opt=paths=source_relative state.proto
